% Generated by roxygen2 (4.0.2): do not edit by hand
\name{callMs}
\alias{callMs}
\title{Function to perform simulation using ms}
\usage{
callMs(opts, dm, subgroup = NA)
}
\arguments{
\item{opts}{The options to pass to ms. Must either be character vector.}

\item{dm}{The demographic model to simulate.}

\item{subgroup}{The subgroup that we are simulating.}
}
\value{
The file containing the output of ms
}
\description{
Function to perform simulation using ms
}

