% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coala_interface.R
\name{create_jaatha_data.segsites_list}
\alias{create_jaatha_data.segsites_list}
\title{Import Population Genetic Data using a list of segregating sites}
\usage{
\method{create_jaatha_data}{segsites_list}(data, model, coala_model,
  trios = NULL, ...)
}
\arguments{
\item{data}{The data to be analysed with Jaatha.
It should be in a format identical to the
simulation results (see \code{\link{create_jaatha_model}}).}

\item{model}{The jaatha model, see \code{\link{create_jaatha_model}}.}

\item{coala_model}{The coala model that was used for creating \code{model}.}

\item{trios}{If you are using locus trios for your anaylsis, then you need
to tell Jaatha with loci should be combined to a trio. To do so,
set this parameter to a list, in which each entry is a numeric
vector of length three containing the indexes of three loci
which will form a trio.}

\item{...}{Currently ignored.}
}
\description{
Import Population Genetic Data using a list of segregating sites
}

