% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coala_interface.R
\name{create_jaatha_model.coalmodel}
\alias{create_jaatha_model.coalmodel}
\title{importFrom coala simulate.coalmodel}
\usage{
\method{create_jaatha_model}{coalmodel}(x, ..., test = TRUE)
}
\description{
importFrom coala simulate.coalmodel
}

