% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaatha_model.R
\name{create_jaatha_model.default}
\alias{create_jaatha_model.default}
\title{For other uappropriate create_jaatha_model will throw an error}
\usage{
\method{create_jaatha_model}{default}(x, ..., scaling_factor = 1, test = TRUE)
}
\arguments{
\item{x}{The primary argument. Can be a function used for simulations,
or a coala model.}

\item{...}{Additional parameters passed on to the dispatch function.}

\item{scaling_factor}{If your model is a down-scaled version of your data,
you can indicated this using this value. The estimated expectation values
are multiplied with this factor before the likelihood is calculated.}

\item{test}{A logical indicating whether a simulation is performed to test
the model.}
}
\description{
For other uappropriate create_jaatha_model will throw an error
}
