% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jaatha_model.R
\name{create_jaatha_model}
\alias{create_jaatha_model}
\title{Specify a model for a Jaatha analysis}
\usage{
create_jaatha_model(x, ..., scaling_factor = 1, test = TRUE)
}
\arguments{
\item{x}{The primary argument. Can be a function used for simulations,
or a coala model.}

\item{...}{Additinoal parameters passed on to the dispatched functions.}

\item{scaling_factor}{If your model is a down-scaled version of your data,
you can indicated this using this value. The estimated expectation values
are multiplied with this factor before the likelihood is calculated.}

\item{test}{A logical indicating whether a simulation is performed to test
the model.}
}
\description{
Specify a model for a Jaatha analysis
}

