% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dm_class.R
\name{dm.addPositiveSelection}
\alias{dm.addPositiveSelection}
\title{Adds positiv selection to a model}
\usage{
dm.addPositiveSelection(dm, min.strength = NA, max.strength = NA,
  parameter = "s", variance = 0, fraction.neutral = 0, population,
  at.time, group = 0)
}
\arguments{
\item{dm}{The demographic model to which mutations should be added}

\item{min.strength}{Minimal strength of selection}

\item{max.strength}{Maximal strength of selection}

\item{parameter}{Instead of creating a new parameter, you can also
set the mutation rate to an expression based on existing
parameters. For example setting this to "theta" will use
an parameter with name theta that you have previously
created. You can also use R expression here, so "2*theta"
or "5*theta+2*tau" (if tau is another parameter) will also
work (also it does not make much sense).}

\item{variance}{Set to a value different from 0 to introduce variation in the
the parameter value for different loci. The
variation follows a gamma distribution with mean equal to
the value provided as \code{parameter}, and variance as given
here. Can also be set to a previously
created parameter, or an expression based on parameters.}

\item{fraction.neutral}{Optionally, a fraction of the loci in the group can
be neutral.}

\item{population}{The populaton in which the allele is selected.}

\item{at.time}{The time at which the selection starts.}

\item{group}{Group of loci for with this feature is added. 0 means that
the feature applies to all groups, and 1 is the default group.
Set to 1 or an greater integer to set this feature only for
the corresponding group of loci.}
}
\description{
Adds positiv selection to a model
}

