% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dm.addSampleSize}
\alias{dm.addSampleSize}
\title{Sets how many individuals from each population are sampled at time 0.}
\usage{
dm.addSampleSize(dm, sample.size)
}
\arguments{
\item{dm}{The demographic model to which recombination events should be added.}

\item{sample.size}{A vector with sample sizes for each population.}
}
\value{
The demographic model with the sample
}
\description{
Sets how many individuals from each population are sampled at time 0.
}

