% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dm.createThetaTauModel}
\alias{dm.createThetaTauModel}
\title{Creates a standard "Theta/Tau" demopraphic model.}
\usage{
dm.createThetaTauModel(sample.sizes, loci.num, seq.length = 1000)
}
\arguments{
\item{sample.sizes}{A numeric vector with the sample sizes of the two pop.sources.}

\item{loci.num}{The number of Loci to simulate.}

\item{seq.length}{For recombination, each locus is assumed to be of this
length}
}
\value{
A Theta/Tau Model
}
\description{
Creates a standard "Theta/Tau" demopraphic model.
}
\examples{
dm <- dm.createThetaTauModel(c(20,25), 100)
dm
}

