% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estimate_log_likelihood.R
\name{estimateLogLikelihood}
\alias{estimateLogLikelihood}
\title{Estimates the likelihood for a parameter combination}
\usage{
estimateLogLikelihood(param, glm_fitted, sum_stats, scaling_factor = 1)
}
\arguments{
\item{param}{The parameter combination. Must be a named vector where the
names are the name of the parameters.}

\item{glm_fitted}{A list of the fitted GLMs, as produced by fitGLM().}

\item{sum_stats}{The summary statistic description, as in jaatha's sum.stats}

\item{scaling_factor}{The scaling factor used for the simulations
slot.}
}
\value{
The estimated log-likelihood
}
\description{
Estimates the likelihood for a parameter combination
}

