% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/find_best_par_in_block.R
\name{findBestParInBlock}
\alias{findBestParInBlock}
\title{Function that uses the fitted GLMs to estimate maximum likelihood
parameters within a block.}
\usage{
findBestParInBlock(block, glm_fitted, sum_stats, scaling_factor = 1)
}
\arguments{
\item{block}{The block we are in.}

\item{glm_fitted}{The fitted GlMs.}

\item{sum_stats}{The observed summary statistics}

\item{scaling_factor}{The scaling factor used for the simulations}
}
\value{
A list with maximum likelihood parameter (\code{est})
  and log-likelihood (\code{score})
}
\description{
Function that uses the fitted GLMs to estimate maximum likelihood
parameters within a block.
}
\author{
Paul Staab & Lisha Mathew
}

