% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_glm.R
\name{fit_glm.Jaatha}
\alias{fit_glm.Jaatha}
\title{Fits a Generalized Linear Model within a block.}
\usage{
\method{fit_glm}{Jaatha}(jaatha, sim_data)
}
\arguments{
\item{jaatha}{A Jaatha Object.}

\item{sim_data}{Simulation data in this block.}
}
\value{
A list containing a list of fitted GLMs for each summary
statistic.
}
\description{
The model describes how the summary statistics depend on the parameters.
}

