% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fitGlm.Jaatha}
\alias{fitGlm.Jaatha}
\title{Fits a Generalized Linear Model within a block.}
\usage{
\method{fitGlm}{Jaatha}(jaatha, sim_data)
}
\arguments{
\item{jaatha}{A Jaatha Object.}

\item{sim_data}{Simulation data in this block.}
}
\value{
A list containing a list of fitted GLMs for each summary
statistic.
}
\description{
The model describes how the summary statistics depend on the parameters.
}

