% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fit_glm.R
\name{fitGlm.Stat_PoiInd}
\alias{fitGlm.Stat_PoiInd}
\title{Fits a GLM for a summary statistics of type "poisson.independent"}
\usage{
\method{fitGlm}{Stat_PoiInd}(sum_stat, sim_data)
}
\arguments{
\item{sum_stat}{Name of the summary statistics}

\item{sim_data}{Results from simulations}
}
\value{
A list of fitted GLMs, one for each function
}
\description{
Fits a GLM for a summary statistics of type "poisson.independent"
}

