% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fit_glm.R
\name{fitGlm.Stat_PoiSmooth}
\alias{fitGlm.Stat_PoiSmooth}
\title{Fits a GLM for a summary statistics of type "poisson.smoothed"}
\usage{
\method{fitGlm}{Stat_PoiSmooth}(sum_stat, sim_data)
}
\arguments{
\item{sum_stat}{Name of the summary statistics}

\item{sim_data}{Results from simulations}
}
\value{
A list with one fitted GLM
}
\description{
Fits a GLM for a summary statistics of type "poisson.smoothed"
}

