% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitPoiSmoothed}
\alias{fitPoiSmoothed}
\title{Fits a GLM for a summary statistics of type "poisson.smoothed"}
\usage{
fitPoiSmoothed(sim.data, sum.stat, weighting, jaatha)
}
\arguments{
\item{sim.data}{Results from simulations}

\item{sum.stat}{Name of the summary statistics}

\item{weighting}{Potentially weights for the simulations.}

\item{jaatha}{A Jaatha Object.}
}
\value{
A list with one fitted GLM
}
\description{
Fits a GLM for a summary statistics of type "poisson.smoothed"
}

