\docType{package}
\name{jaatha-package}
\alias{jaatha-package}
\title{Fast estimation of demographic parameters}
\description{
  Fast estimation of demographic parameters for two species
}
\details{
  Roxygen is a Doxygen-like documentation system for R;
  allowing in-source specification of Rd files, collation
  and namespace directives.

  If you have existing Rd files, check out the
  \code{Rd2roxygen} package for a convenient way of
  converting Rd files to roxygen comments.

  \tabular{ll}{ Package: \tab jaatha\cr Type: \tab
  Package\cr Version: \tab 0.9.1\cr Date: \tab
  2012-05-10\cr License: \tab GPL (>= 2)\cr LazyLoad: \tab
  yes\cr }
}
\author{
  Lisha Naduvilezhath
  \email{lisha@biologie.uni-muenchen.de}, Paul R. Staab
  \email{staab@biologie.uni-muenchen.de} and Dirk Metzler
  \email{metzler@biologie.uni-muenchen.de}

  Maintainer: Paul R. Staab
  \email{staab@biologie.uni-muenchen.de}
}
\keyword{package}

