% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Jaatha.getCIsFromLogs}
\alias{Jaatha.getCIsFromLogs}
\title{Function for calculating bootstrap confidence intervals from logs of a
previous run of Jaatha.confidenceIntervals.}
\usage{
Jaatha.getCIsFromLogs(jaatha, conf_level = 0.95, log_folder)
}
\arguments{
\item{jaatha}{The Jaatha object that was used when calling
Jaatha.confidenceIntervals.}

\item{conf_level}{The intended confidence level of the interval. The actual
level can vary slightly.}

\item{log_folder}{The folder with logs from the previous run. Just use the
folder that was given as 'log.folder' argument there.}
}
\value{
The Jaatha Object with confidence intervals included.
}
\description{
Function for calculating bootstrap confidence intervals from logs of a
previous run of Jaatha.confidenceIntervals.
}

