% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/likelihood_table.R
\name{Jaatha.getLikelihoods}
\alias{Jaatha.getLikelihoods}
\title{Gives the best estimates after a Jaatha search}
\usage{
Jaatha.getLikelihoods(jaatha, max_rows = NULL, initial_search = FALSE)
}
\arguments{
\item{jaatha}{The Jaatha options}

\item{initial_search}{If \code{TRUE}, then the estimates for the blocks in
the initial search will be reported.}

\item{max.entries}{If given, no more than this number of entries will be
returned.}
}
\value{
A matrix with log composite likelihoods and parameters of The
best estimates
}
\description{
This method extracts the best estimates with log composite likelihood
vales from an Jaatha object.
}

