\name{Jaatha.pickBestStartPoints}
\alias{Jaatha.pickBestStartPoints}
\title{Pick best starting positions}
\usage{
  Jaatha.pickBestStartPoints(blocks, best)
}
\arguments{
  \item{blocks}{A list of starting positions (typically
  from \code{\link{Jaatha.initialSearch}})}

  \item{best}{The number of best starting positions to
  keep}
}
\value{
  A list with the best starting positions
}
\description{
  This functions picks the best starting positions out of a
  list of starting positions.
}

