\name{Jaatha.printLikelihoods}
\alias{Jaatha.printLikelihoods}
\title{Gives the best estimates after a Jaatha search}
\usage{
  Jaatha.printLikelihoods(jObject)
}
\arguments{
  \item{jObject}{The Jaatha options}
}
\value{
  A matrix with log composite likelihoods and parameters of
  The best estimates
}
\description{
  This method extracts the best estimates with log
  composite likelihood vales from an Jaatha object.
}

