\name{Jaatha.printStartPoints}
\alias{Jaatha.printStartPoints}
\title{Print Start points}
\usage{
  Jaatha.printStartPoints(jObject, startPoints)
}
\arguments{
  \item{jObject}{The Jaatha options}

  \item{startPoints}{The list of startPoints which will be
  printed}
}
\value{
  a matrix with score and parameters of each start point
}
\description{
  Method to print the start Points given by an initial
  Jaatha search sorted by score.
}

