% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Jaatha.setSeqgenExecutable}
\alias{Jaatha.setSeqgenExecutable}
\title{Set the path to the executable for seqgen}
\usage{
Jaatha.setSeqgenExecutable(seqgen.exe)
}
\arguments{
\item{seqgen.exe}{Path to seqgen's executable.}
}
\description{
Set the path to the executable for seqgen
}

