\name{Jaatha.setSimulBandsExecutable}
\alias{Jaatha.setSimulBandsExecutable}
\title{Set the path to the executable for seqgen}
\usage{
Jaatha.setSimulBandsExecutable(seqgen.exe)
}
\arguments{
  \item{seqgen.exe}{Path to seqgen's executable.}
}
\description{
Set the path to the executable for seqgen
}

