% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setCores}
\alias{setCores}
\title{Function to set the number of cores that Jaatha uses for simulations}
\usage{
setCores(jaatha, cores = 1)
}
\arguments{
\item{jaatha}{The jaatha object for which we set the number of cores}

\item{cores}{The number of cores to use. On Windows, only one core can be
used.}
}
\value{
The jaatha object with set number of cores
}
\description{
Jaatha can distribute the simulation on multiple cores if the operating
system supports p-threads (parrallel::mclapply is used internally). This
function sets the number of cores that should be used.
}

