% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popgenome_interface.R
\name{create_jaatha_data.GENOME}
\alias{create_jaatha_data.GENOME}
\title{Use PopGenome Data for Jaatha}
\usage{
\method{create_jaatha_data}{GENOME}(data, model, coala_model, trios = NULL,
  only_synonymous = FALSE, ...)
}
\arguments{
\item{data}{The \code{GENOME} data from \pkg{PopGenome}.}

\item{model}{The jaatha model, see \code{\link{create_jaatha_model}}.}

\item{coala_model}{The coala model that was provided to 
\code{\link{create_jaatha_model}}.}

\item{trios}{An optional argument that will be passed on to 
\code{coala::calc_sumstats_from_data} and is documented there.}

\item{only_synonymous}{Only use synonymous SNP if set to \code{TRUE}. 
This requires that \pkg{PopGenome} knows where coding regions are., e.g.
by using gff files.}

\item{...}{Currently ignored.}
}
\description{
This allows to use genetic data that was imported with the package
\pkg{PopGenome} to be analysed with Jaatha. As this relies on \pkg{coala}
to calculate the summary statistics, you need to use a model that was created
from a coala model.
}

