% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jaatha_data.R
\name{create_jaatha_data}
\alias{create_jaatha_data}
\title{Prepare data for use in jaatha}
\usage{
create_jaatha_data(data, model, ...)
}
\arguments{
\item{data}{The data that is imported}

\item{model}{The model used for the estimation.
See \code{\link{create_jaatha_model}}.}

\item{...}{Additinal arguments that might be used in a dispatched function}
}
\description{
Prepare data for use in jaatha
}

