\name{dm.getLociNumber}
\alias{dm.getLociNumber}
\title{Gets how many loci belong to a group of loci}
\usage{
dm.getLociNumber(dm, group = NULL)
}
\arguments{
  \item{dm}{The Demographic Model}

  \item{group}{The group for which we get the number of
  loci}
}
\value{
The number of loci in the group
}
\description{
Gets how many loci belong to a group of loci
}

