% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dm.getLociNumber}
\alias{dm.getLociNumber}
\title{Gets how many loci belong to a group of loci}
\usage{
dm.getLociNumber(dm, group = 1)
}
\arguments{
\item{dm}{The Demographic Model}

\item{group}{The group for which we get the number of loci. Defaults to
the first group.}
}
\value{
The number of loci in the group
}
\description{
Gets how many loci belong to a group of loci
}

