% Generated by roxygen2 (4.0.2): do not edit by hand
\name{estimateLogLikelihood}
\alias{estimateLogLikelihood}
\title{Estimates the likelihood for a parameter combination}
\usage{
estimateLogLikelihood(param, glm.fitted, sum.stats)
}
\arguments{
\item{param}{The parameter combination. Must be a named vector where the
names are the name of the parameters.}

\item{glm.fitted}{A list of the fitted GLMs, as produced by fitGLM().}

\item{sum.stats}{The summary statistic description, as in jaatha's sum.stats
slot.}
}
\value{
The estimated log-likelihood
}
\description{
Estimates the likelihood for a parameter combination
}

