% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/find_best_par_in_block.R
\name{findBestParInBlock}
\alias{findBestParInBlock}
\title{Functions that uses the fitted GLMs to estimate the parameters that have the
highest likelihood.}
\usage{
findBestParInBlock(block, glm.fitted, sum.stats, scaling_factor = 1)
}
\arguments{
\item{block}{The block we are in.}

\item{glm.fitted}{The fitted GlMs.}

\item{sum.stats}{The observed summary statistics}

\item{scaling_factor}{The scaling factor used for the simulations}
}
\value{
A list with maximum likelihood parameter (est) and log-likelihood
(score)
}
\description{
Functions that uses the fitted GLMs to estimate the parameters that have the
highest likelihood.
}

