% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jaatha.R
\docType{package}
\name{jaatha-package}
\alias{jaatha-package}
\title{Fast estimation of demographic parameters}
\description{
Jaatha is a composite maximum likelihood method to estimate parameters
of a speciation model of closely related biological species out of SNP
data.
}
\author{
Lisha Naduvilezhath \email{lisha (at) biologie.uni-muenchen.de},
Paul R. Staab \email{staab (at) biologie.uni-muenchen.de} and
Dirk Metzler \email{metzler (at) biologie.uni-muenchen.de}

Maintainer: Paul R. Staab \email{staab (at) biologie.uni-muenchen.de}
}
\keyword{package}

