% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaatha.R
\docType{package}
\name{jaatha-package}
\alias{jaatha-package}
\title{Jaatha: Simulation-based maximum likelihood parameter estimation}
\description{
Jaatha is a simulation based maximum likelihood parameter estimation
methods.
}
\keyword{package}

