\name{Jaatha.getStartingPoints}
\alias{Jaatha.getStartingPoints}
\title{Print Start points}
\usage{
  Jaatha.getStartingPoints(jObject)
}
\arguments{
  \item{jObject}{The Jaatha options}
}
\value{
  a matrix with score and parameters of each start point
}
\description{
  Method to print the start Points given by an initial
  Jaatha search sorted by score.
}

