% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jaatha.R
\name{Jaatha.initialize}
\alias{Jaatha.initialize}
\title{Initialization of a Jaatha estimation for population genetics}
\usage{
Jaatha.initialize(data, model, cores = 1, scaling_factor = 1,
  smoothing = FALSE, only_synonymous = FALSE)
}
\arguments{
\item{data}{The observed data. Jaatha can use data imported with package
\pkg{PopGenome}. Please refer the to the vignette
"The Jaatha HowTo" for more information.}

\item{model}{The demographic model to use}

\item{cores}{The number of cores to use in parallel. If 0, it tries to
guess the number of available cores and use them all.}

\item{scaling_factor}{You can use this option if you have a large dataset. If
so, Jaatha only simulates only a fraction 1/scaling_factor of the
dataset and interpolates the missing data.}

\item{smoothing}{If set to true, Jaatha uses a different way to summaries the
JSFS. Instead of binning certain areas, and fitting a glm per
area, only one glm is fitted for the complete JSFS, and the
position of the different entries is treated as a model
parameter. This feature is still experimental and not
recommended for productive use at the moment.}

\item{only_synonymous}{Only use synonymous SNP if set to \code{TRUE}. Requires
to provided \code{data} as a PopGenome "GENOME" object.}
}
\value{
A S4-Object of type jaatha containing the settings
}
\description{
This function sets the basic parameters for an analysis with
Jaatha and is the first step for each application of it.
}

