\name{normalize}
\alias{normalize}
\title{Convert parameters from the natural scale into Jaatha's internal scale}
\usage{
normalize(value, jaatha)
}
\arguments{
  \item{value}{A parameter combination in this natural
  scale}

  \item{jaatha}{The current jaatha object}
}
\value{
The parameter from value, converted into Jaatha's interval
0-1 scale
}
\description{
Convert parameters from the natural scale into Jaatha's
internal scale
}

